Add files via upload
The goal here is to classify these images, and other several images of bales that we receive on a daily basis on types 1, 2 and 3.
I think that first, we need to establish a data bank of several images of type 1, 2 and 3, so that the tool starts to learn and classiy bales based on some parameters. Comments and feedback are always welcome.

Knowledge of image detection/machine learning = just started reading about this.